Move exported widget objects to @vcd/widget-object
