+++ projects/widget-object/src/index.ts
@@ -2,4 +2,4 @@
  * Copyright 2021 VMware, Inc.
  * SPDX-License-Identifier: BSD-2-Clause
  */
-export * from './lib/widget-object';
+export * from './lib/';
