+++ projects/widget-object/src/lib/index.ts
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-export { BaseWidgetObject, FindableWidget } from './widget-object';
 export { CypressWidgetObjectFinder, FindCypressWidgetOptions } from './cypress/cypress-widget-finder';
 export { AngularWidgetObjectFinder } from './angular/angular-widget-finder';
 export { TestElement } from './angular/angular-widget-object-element';
-export { FindElementOptions } from './widget-object';
+export { FindElementOptions, BaseWidgetObject, FindableWidget, WidgetObjectElement } from './widget-object';
+export { ClrDatagridWidgetObject } from './datagrid/datagrid.wo';
+export { VcdDatagridWidgetObject } from './datagrid/vcd-datagrid.wo';
